requirements.txt
matplotlib==3.2.1
numpy==1.18.2
Pillow==6.2.1
scikit-learn==0.22
scipy==1.4.1
jupyter==1.0.0      
jupyter-client==6.1.3      
jupyter-console==6.1.0      
jupyter-core==4.6.3   
